chore: refactor OpenAI-compatible providers
