+++ core/src/lib.rs
@@ -56,6 +56,7 @@ pub mod providers {
     pub mod anthropic;
     pub mod deepseek;
     pub mod google_ai_studio;
+    pub mod openai_compatible_helpers;
     pub mod togetherai;
 }
 pub mod http {
