+++ core/src/providers/deepseek.rs
@@ -1,12 +1,9 @@
-use crate::providers::chat_messages::{AssistantChatMessage, ChatMessage};
+use std::sync::Arc;
+
+use crate::providers::chat_messages::ChatMessage;
 use crate::providers::embedder::Embedder;
 use crate::providers::llm::ChatFunction;
-use crate::providers::llm::{LLMChatGeneration, LLMGeneration, LLMTokenUsage, LLM};
-use crate::providers::openai::{
-    chat_completion, logprobs_from_choices, streamed_chat_completion, to_openai_messages,
-    OpenAIChatMessage, OpenAIChatMessageContent, OpenAIContentBlock, OpenAITextContent,
-    OpenAITextContentType, OpenAITool, OpenAIToolChoice,
-};
+use crate::providers::llm::{LLMChatGeneration, LLMGeneration, LLM};
 use crate::providers::provider::{Provider, ProviderID};
 use crate::providers::tiktoken::tiktoken::{batch_tokenize_async, o200k_base_singleton, CoreBPE};
 use crate::providers::tiktoken::tiktoken::{decode_async, encode_async};
@@ -18,9 +15,11 @@ use async_trait::async_trait;
 use hyper::Uri;
 use parking_lot::RwLock;
 use serde_json::Value;
-use std::str::FromStr;
-use std::sync::Arc;
 use tokio::sync::mpsc::UnboundedSender;
+
+use super::openai_compatible_helpers::{
+    openai_compatible_chat_completion, TransformSystemMessages,
+};
 
 pub struct DeepseekLLM {
     id: String,
@@ -115,7 +114,7 @@ impl LLM for DeepseekLLM {
         top_p: Option<f32>,
         n: usize,
         stop: &Vec<String>,
-        mut max_tokens: Option<i32>,
+        max_tokens: Option<i32>,
         presence_penalty: Option<f32>,
         frequency_penalty: Option<f32>,
         logprobs: Option<bool>,
@@ -123,138 +122,29 @@ impl LLM for DeepseekLLM {
         _extras: Option<Value>,
         event_sender: Option<UnboundedSender<Value>>,
     ) -> Result<LLMChatGeneration> {
-        if let Some(m) = max_tokens {
-            if m == -1 {
-                max_tokens = None;
-            }
-        }
-
-        let tool_choice = match function_call.as_ref() {
-            Some(fc) => Some(OpenAIToolChoice::from_str(fc)?),
-            None => None,
-        };
-
-        let tools = functions
-            .iter()
-            .map(OpenAITool::try_from)
-            .collect::<Result<Vec<OpenAITool>, _>>()?;
-
-        // Deepseek doesn't work with the new chat message content format.
-        // We have to modify the messages contents to use the "String" format.
-        let openai_messages = to_openai_messages(messages, &self.id)?
-            .into_iter()
-            .filter_map(|m| match m.content {
-                None => Some(m),
-                Some(OpenAIChatMessageContent::String(_)) => Some(m),
-                Some(OpenAIChatMessageContent::Structured(contents)) => {
-                    // Find the first text content, and use it to make a string content.
-                    let content = contents.into_iter().find_map(|c| match c {
-                        OpenAIContentBlock::TextContent(OpenAITextContent {
-                            r#type: OpenAITextContentType::Text,
-                            text,
-                            ..
-                        }) => Some(OpenAIChatMessageContent::String(text)),
-                        _ => None,
-                    });
-
-                    Some(OpenAIChatMessage {
-                        role: m.role,
-                        name: m.name,
-                        tool_call_id: m.tool_call_id,
-                        tool_calls: m.tool_calls,
-                        content,
-                    })
-                }
-            })
-            .collect::<Vec<_>>();
-
-        let is_streaming = event_sender.is_some();
-
-        let (c, request_id) = if is_streaming {
-            streamed_chat_completion(
-                self.chat_uri()?,
-                self.api_key.clone().unwrap(),
-                None,
-                Some(self.id.clone()),
-                &openai_messages,
-                tools,
-                tool_choice,
-                temperature,
-                match top_p {
-                    Some(t) => t,
-                    None => 1.0,
-                },
-                n,
-                stop,
-                max_tokens,
-                match presence_penalty {
-                    Some(p) => p,
-                    None => 0.0,
-                },
-                match frequency_penalty {
-                    Some(f) => f,
-                    None => 0.0,
-                },
-                None,
-                None,
-                logprobs,
-                top_logprobs,
-                None,
-                event_sender.clone(),
-            )
-            .await?
-        } else {
-            chat_completion(
-                self.chat_uri()?,
-                self.api_key.clone().unwrap(),
-                None,
-                Some(self.id.clone()),
-                &openai_messages,
-                tools,
-                tool_choice,
-                temperature,
-                match top_p {
-                    Some(t) => t,
-                    None => 1.0,
-                },
-                n,
-                stop,
-                max_tokens,
-                match presence_penalty {
-                    Some(p) => p,
-                    None => 0.0,
-                },
-                match frequency_penalty {
-                    Some(f) => f,
-                    None => 0.0,
-                },
-                None,
-                None,
-                logprobs,
-                top_logprobs,
-                None,
-            )
-            .await?
-        };
-
-        assert!(c.choices.len() > 0);
-
-        Ok(LLMChatGeneration {
-            created: utils::now(),
-            provider: ProviderID::Deepseek.to_string(),
-            model: self.id.clone(),
-            completions: c
-                .choices
-                .iter()
-                .map(|c| AssistantChatMessage::try_from(&c.message))
-                .collect::<Result<Vec<_>>>()?,
-            usage: c.usage.map(|usage| LLMTokenUsage {
-                prompt_tokens: usage.prompt_tokens,
-                completion_tokens: usage.completion_tokens.unwrap_or(0),
-            }),
-            provider_request_id: request_id,
-            logprobs: logprobs_from_choices(&c.choices),
-        })
+        openai_compatible_chat_completion(
+            self.chat_uri()?,
+            self.id.clone(),
+            self.api_key.clone().unwrap(),
+            messages,
+            functions,
+            function_call,
+            temperature,
+            top_p,
+            n,
+            stop,
+            max_tokens,
+            presence_penalty,
+            frequency_penalty,
+            logprobs,
+            top_logprobs,
+            None,
+            event_sender,
+            false, // don't disable provider streaming
+            TransformSystemMessages::Keep,
+            "DeepSeek".to_string(),
+        )
+        .await
     }
 }
